refactor(primitive): rename primitives.py to primitive.py
