+++ plb/engine/controller/robot_controller.py
@@ -9,7 +9,7 @@ from yacs.config import CfgNode as CN
 from plb.config.utils import make_cls_config
 from .controller import Controller
 from plb.urdfpy import DiffRobot, Robot, Collision, DEVICE
-from plb.engine.primitive.primitives import Box, Sphere, Cylinder, Primitive
+from plb.engine.primitive.primitive import Box, Sphere, Cylinder, Primitive
 from .primitive_controller import PrimitivesController
 
 ROBOT_LINK_DOF = 7
